#7: Setup Typescript on the server
Also added a couple new scripts for building and dev
