+++ package.json
@@ -6,9 +6,12 @@
     "server"
   ],
   "scripts": {
-    "client": "yarn workspace client build",
-    "server": "yarn workspace server start",
-    "start": "concurrently --kill-others-on-fail \"yarn server\" \"yarn client\""
+    "dev:client": "yarn workspace client dev",
+    "dev:server": "yarn workspace server dev",
+    "dev": "concurrently --kill-others-on-fail \"yarn dev:client\" \"yarn dev:server\"",
+    "prod:client": "yarn workspace client build",
+    "prod:server": "yarn workspace server prod",
+    "prod": "concurrently --kill-others-on-fail \"yarn prod:client\" \"yarn prod:server\""
   },
   "dependencies": {
     "concurrently": "^7.0.0"
