+++ server/package.json
@@ -3,12 +3,21 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js"
+    "dev": "NODE_ENV=dev ts-node-dev ./src/server.ts",
+    "prod": "tsc && pm2 stop server && NODE_ENV=production pm2 start ./dist/server.js --name server"
   },
   "license": "MIT",
   "dependencies": {
     "cors": "^2.8.5",
     "express": "4.17.1",
     "nodemon": "2.0.3"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/node": "^17.0.9",
+    "pm2": "^5.1.2",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "^4.5.4"
   }
 }
